Improve security defenses at compile level
This patch adds compiler flags to improve the security of STX code.
At the same time, these patches fix warnings and compilation
issues generated by new security flags.
The flags added are divided in the following categories:

Stack execution protection:              LDFLAGS="-z noexecstack"
Data relocation and protection (RELRO):  LDLFAGS="-z relro -z now"
Stack-based Buffer Overrun Detection:    CFLAGS=”-fstack-protector-strong”
                                            if using GCC 4.9 or newer
Fortify source:                          CFLAGS="-O2 -D_FORTIFY_SOURCE=2"
Format string vulnerabilities:           CFLAGS="-Wformat -Wformat-security"

             Victor Rodriguez <vm.rod25@gmail.com>
Sugested-by: Victor Rodriguez <vm.rod25@gmail.com>
             Erich Cordoba <erich.cordoba.malibran@intel.com>

+++ service-mgmt/sm-common-1.0.0/src/sm_utils.c
@@ -36,7 +36,11 @@ bool sm_utils_process_running( const char* pid_filename )
     } else {
         int pid;
 
-        fscanf( fp, "%d", &pid );
+        if( fscanf( fp, "%d", &pid ) != 1 )
+        {
+            running = false;
+            return(running);
+        }
         if( pid != getpid() )
         {
             result = kill( pid, 0 );
